refactor(portable_dropbot_ui): split package into MVC folders
Organize the flat package into models/, views/, controllers/ and
message_handlers/ (dramatiq topic -> model updates), keeping consts,
plugin and the dock panes at the root. Rename the main pane
controller to PortableDropbotStatusAndControlsController to match
its model and message-handler siblings, its module to
status_controls_pane_controller.py, and dock_pane.py to
dock_panes.py (it defines all six panes). All moves are git renames;
only import paths changed.
