+++ portable_dropbot_status_and_controls/models/calibration_model.py
@@ -6,7 +6,7 @@ from portable_dropbot_controller.consts import (
 )
 from template_status_and_controls.base_model import BaseStatusModel
 
-from .consts import PORTABLE_DROPBOT_IMAGE
+from ..consts import PORTABLE_DROPBOT_IMAGE
 
 
 class PortableDropbotCalibrationModel(BaseStatusModel):
