+++ portable_dropbot_status_and_controls/models/model.py
@@ -11,7 +11,7 @@ from portable_dropbot_controller.consts import (
 )
 from template_status_and_controls.base_model import BaseStatusModel
 
-from .consts import PORTABLE_DROPBOT_IMAGE
+from ..consts import PORTABLE_DROPBOT_IMAGE
 
 
 class PortableDropbotStatusAndControlsModel(BaseStatusModel):
