+++ portable_dropbot_status_and_controls/models/temp_lighting_model.py
@@ -7,7 +7,7 @@ from portable_dropbot_controller.consts import (
 )
 from template_status_and_controls.base_model import BaseStatusModel
 
-from .consts import PORTABLE_DROPBOT_IMAGE
+from ..consts import PORTABLE_DROPBOT_IMAGE
 
 
 class PortableDropbotTempLightingModel(BaseStatusModel):
